spread common.py among dataset modules
